requirements.txt
# ComplexUncertainty 项目依赖

# Web框架
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# HTTP客户端
requests>=2.28.0

# 深度学习框架
torch>=2.0.0
transformers>=4.30.0

# 数值计算
numpy>=1.24.0

# 配置文件解析
pyyaml>=6.0

# 测试
pytest>=7.0.0

# 可选：数据可视化
# matplotlib>=3.7.0
# seaborn>=0.12.0
